api-key: remove create command as it should be handled by the dashboard only
The create API key command have been removed.
API key will tend to be the only way to login and use the CLI.
This behavior is similar to what you usually see on other CLI like gcloud. A user should provide an API key to use the CLI.

We will encourage users to generate and manage their API key using Cycloid dashboard and not by the CLI itself.