code: update component create/update logic
Component creation & configuration was split into 2 parts on the API, but we also added a PUT that combines these two into one call.

Updated the code to reflect these changes.

Breaking: components update command will now fail if stack-ref isn't provided for non-existing component. It's a step forward to combine the create & update commands into one, so stack-ref is necessary in case the component doesn't exist and needs to be created. If it exists, it's updated otherwise